refactor: decompose polish modules under the 1k-line bar
Extract reveal/reactions/share/settings/scores and prompt packs so main.ts stays under 1140 lines and engine ownership stays clearer.
