+++ client/package.json
@@ -18,6 +18,7 @@
   "devDependencies": {
     "@playwright/test": "^1.60.0",
     "@types/qrcode": "^1.5.5",
+    "happy-dom": "^20.10.6",
     "typescript": "^6.0.3",
     "vite": "^8.0.13",
     "vitest": "^4.1.7"
