+++ client/src/styles/results-reveal.css
@@ -1,0 +1,24 @@
+.action-hint {
+  margin: 0;
+  color: var(--muted-strong);
+  font-size: 0.98rem;
+  font-weight: 850;
+}
+
+.display .room-panel { position: relative; }
+.display .room-code-wrap,
+.display .qr-stage { position: relative; z-index: 1; }
+.display .settings-panel { opacity: 0.96; }
+
+.results-panel .reveal-stage {
+  opacity: 0;
+  transform: translateY(10px);
+  transition: opacity 280ms ease, transform 280ms ease;
+  pointer-events: none;
+}
+.results-panel .reveal-stage.is-visible {
+  opacity: 1;
+  transform: none;
+  pointer-events: auto;
+}
+.results-panel[data-reveal-stage='complete'] .confetti { opacity: 1; }
